+++ src/shared/interfaces/paired-term.interface.ts
@@ -1,0 +1,9 @@
+import { ApiProperty } from "@nestjs/swagger";
+
+export class PairedTerm {
+  @ApiProperty()
+  technology: string;
+
+  @ApiProperty()
+  pairings: string[];
+}